notification-consumer/requirements.txt
Flask
Flask-SQLAlchemy
flask-marshmallow
marshmallow_sqlalchemy
marshmallow
pymysql
cryptography
pytest
Flask-Testing
python-dotenv
pika
flask_restplus
celery
firebase_admin
